requirements.txt

Flask==2.3.3
python-magic==0.4.27
fpdf==1.7.2
colorama==0.4.6
frida==16.0.11
frida-tools==12.1.1
loguru==0.6.0
prompt-toolkit==3.0.38
Pygments==2.14.0
typing_extensions==4.5.0
wcwidth==0.2.6
reflutter